+++ clang/test/Misc/target-invalid-cpu-note/amdgcn.c
@@ -78,4 +78,5 @@
 // CHECK-SAME: {{^}}, gfx11-generic
 // CHECK-SAME: {{^}}, gfx12-generic
 // CHECK-SAME: {{^}}, gfx9-4-generic
+// CHECK-SAME: {{^}}, gfx12-5-generic
 // CHECK-SAME: {{$}}