+++ clang/test/Misc/target-invalid-cpu-note/nvptx.c
@@ -93,6 +93,7 @@
 // CHECK-SAME: {{^}}, gfx1201
 // CHECK-SAME: {{^}}, gfx1250
 // CHECK-SAME: {{^}}, gfx1251
+// CHECK-SAME: {{^}}, gfx12-5-generic
 // CHECK-SAME: {{^}}, gfx1310
 // CHECK-SAME: {{^}}, amdgcnspirv
 // CHECK-SAME: {{$}}